refactor: remove old text input
