+++ frontend/components/form/FormPasswordInput.vue
@@ -1,22 +1,27 @@
 <!-- SPDX-License-Identifier: AGPL-3.0-or-later -->
 <template>
-  <FormTextInput2 :hasError="hasError" :type="isPassword ? 'password' : 'text'">
+  <FormTextInput
+    :id="id"
+    :hasError="hasError"
+    :type="isPassword ? 'password' : 'text'"
+  >
     <template #icons>
       <slot name="icons"></slot>
-      <button @click="changeInputType">
+      <button @click="changeInputType" :id="`${id}-show-password`">
         <Icon
           :name="isPassword ? IconMap.VISIBLE : IconMap.HIDDEN"
           size="1.4em"
         />
       </button>
     </template>
-  </FormTextInput2>
+  </FormTextInput>
 </template>
 
 <script setup lang="ts">
 import { IconMap } from "~/types/icon-map";
 
 export interface Props {
+  id: string;
   hasError?: boolean;
 }
