+++ frontend/pages/auth/set-password.vue
@@ -3,49 +3,53 @@
   <div class="px-4 sm:px-6 md:px-8 xl:px-24 2xl:px-36">
     <form class="space-y-4">
       <FormTextInput
-        @update:model-value="userNameValue = $event"
-        :placeholder="$t('i18n.pages.auth._global.enter_user_name')"
-        :model-value="userNameValue"
+        @input="userName = $event.target.value"
+        id="set-password-username"
+        :value="userName"
+        :placeholder="$t('i18n.pages.auth._global.enter_a_user_name')"
+        :data-testid="$t('i18n.pages.auth._global.enter_a_user_name')"
       />
-      <FormTextInput
-        @update:model-value="passwordValue = $event"
-        @input="checkRules"
-        @blurred="
-          isBlurred = true;
-          isFocused = false;
-        "
-        @focused="
-          isFocused = true;
-          isBlurred = false;
-        "
+      <FormPasswordInput
+        @input="handlePasswordInput"
+        @blur="isPasswordFocused = false"
+        @focus="isPasswordFocused = true"
+        id="set-password-password"
+        :value="password"
         :placeholder="$t('i18n._global.enter_password')"
-        :is-icon-visible="true"
-        input-type="password"
-        :model-value="passwordValue"
-        :icons="[IconMap.VISIBLE]"
-        :error="!isAllRulesValid && isBlurred"
+        :data-testid="$t('i18n._global.enter_password')"
+        :hasError="showPasswordError.border"
       />
-      <IndicatorPasswordStrength :password-value="passwordValue" />
+      <IndicatorPasswordStrength :password-value="password" />
       <TooltipPasswordRequirements
-        v-if="
-          !!passwordValue?.length &&
-          !isAllRulesValid &&
-          (!isBlurred || isFocused)
-        "
+        v-if="showPasswordError.tooltip"
         :rules="rules"
       />
-      <FormTextInput
-        @update:model-value="confirmPasswordValue = $event"
+      <FormPasswordInput
+        @input="confirmPassword = $event.target.value"
+        id="set-password-confirm-password"
+        :value="confirmPassword"
         :placeholder="$t('i18n._global.repeat_password')"
-        :is-icon-visible="true"
-        input-type="password"
-        :model-value="confirmPasswordValue"
-        :icons="
-          isPasswordMatch(passwordValue, confirmPasswordValue)
-            ? [IconMap.CHECK, IconMap.VISIBLE]
-            : [IconMap.X_LG, IconMap.VISIBLE]
-        "
-      />
+        :data-testid="$t('i18n._global.repeat_password')"
+      >
+        <template #icons>
+          <span>
+            <Icon
+              v-if="doPasswordsMatch"
+              :name="IconMap.CHECK"
+              size="1.2em"
+              color="#3BA55C"
+              data-testid="extra-icon"
+            />
+            <Icon
+              v-else
+              :name="IconMap.X_LG"
+              size="1.2em"
+              color="#BA3D3B"
+              data-testid="extra-icon"
+            />
+          </span>
+        </template>
+      </FormPasswordInput>
       <div class="pt-4">
         <BtnAction
           class="flex max-h-[48px] items-center justify-center truncate md:max-h-[40px]"
@@ -61,13 +65,30 @@
 <script setup lang="ts">
 import { IconMap } from "~/types/icon-map";
 
-const userNameValue = ref("");
-const passwordValue = ref("");
-const confirmPasswordValue = ref("");
-
-const isBlurred = ref(false);
-const isFocused = ref(false);
+const userName = ref("");
+const password = ref("");
+const confirmPassword = ref("");
+const isPasswordFocused = ref(false);
 
 const { rules, isAllRulesValid, checkRules, isPasswordMatch } =
   usePasswordRules();
+
+const doPasswordsMatch = computed<boolean>(() =>
+  isPasswordMatch(password.value, confirmPassword.value)
+);
+
+const showPasswordError = computed<{ border: boolean; tooltip: boolean }>(
+  () => {
+    const error = password.value.length > 0 && !isAllRulesValid.value;
+    return {
+      border: !isPasswordFocused.value && error,
+      tooltip: isPasswordFocused.value && error,
+    };
+  }
+);
+
+const handlePasswordInput = (event: Event & { target: HTMLInputElement }) => {
+  password.value = event.target.value;
+  checkRules(event);
+};
 </script>
