+++ frontend/test-e2e/page-objects/SignInPage.ts
@@ -7,9 +7,9 @@ export const newSignInPage = (page: Page) => ({
   signUpLink: page.getByRole("link", {
     name: new RegExp(getEnglishText("i18n._global.sign_up_aria_label"), "i"),
   }),
-  usernameInput: page.locator("#sign-in-username input"),
-  passwordInput: page.locator("#sign-in-password input"),
-  showPasswordToggle: page.locator("#sign-in-password span[role='button']"),
+  usernameInput: page.locator("#sign-in-username"),
+  passwordInput: page.locator("#sign-in-password"),
+  showPasswordToggle: page.locator("#sign-in-password-show-password"),
   captcha: page.locator("#sign-in-captcha"),
 
   // Currently a button but should be a link
